Improvement to quick casting
Does not require a weapon to be equipped for either "ready magic" key or quickkeys.  Leaves "equipped weapon" state untouched in either case.

Only remaining issue is that animations are not smooth; this is probably out of scope to fix as it might require custom animations particularly for the equipped weapon -> cast spell (with weapon in hand) -> equipped weapon sequence.