More work on improved spell casting
Key remaining issue: when accounting for the amount of magicka a player has available in calculating chance of success, it uses the amount after the player cast the spell, rendering the calculation inaccurate when the amount of available magicka is a factor.

There are also some minor rounding issues that need to be fixed.