More fixes for spellcasting overhaul
- Spell chance success always rounds up (this is more consistent with vanilla mechanics as well)
- Fatigue does not incorrectly skew projected cost / chance for spell crafting

Remaining issues: Presentation of spell cost (both for spell crafting and in spell list) when chance < 0, either due to magicka level or skill.  Should display the minimum magicka required to have a chance of successfully casting the spell -- not the players available magicka, or a higher cost due to low skill level.  This would be either the cost of casting it when the player reaches the threshold where that becomes possible if limited by skill level, or the base cost of the spell if limited by magicka.